requirements.txt
aiogram==3.3.0
openai==1.6.1
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.2.0
python-magic==0.4.27
pdf2image==1.17.0
pytesseract==0.3.10
httpx==0.24.1 